migration viser ikke ændring af decimal til int

+++ EFCore/Validators/NonNegativeValidator.cs
@@ -6,7 +6,7 @@ namespace ExperienceAPI.Validators
     {
         protected override ValidationResult IsValid(object value, ValidationContext validationContext)
         {
-            if (value is decimal price && price < 0)
+            if (value is int price && price < 0)
             {
                 return new ValidationResult("Price cannot and must not negative. try again lil bro");
             }
